attached update.save_out to DirectImageGuide
